Rename vscodeTypeChatModel to getTypeChatLanguageModel
